+++ app/controllers/ai_credentials_controller.rb
@@ -4,6 +4,14 @@ class AiCredentialsController < CredentialsController
 
   private
 
+  def credential_params
+    super.merge(params.require(:ai_credential).permit(:default_model))
+  end
+
+  def updated_credential_attrs(credential_updates:)
+    super.merge(credential_params.slice(:default_model))
+  end
+
   def default_provider
     LlmProvider.names.first
   end